Refactor route permission check
