+++ timApp/peerreview/util/groups.py
@@ -1,18 +1,7 @@
-"""
-Short description of Python module
-"""
-
-__authors__ = ["Mika Lehtinen", "Simo Lehtinen", "Alexander Södergård"]
-__license__ = "MIT"
-__date__ = "29.5.2022"
-
 from collections import defaultdict
 from datetime import datetime
 from random import shuffle
 from typing import DefaultDict
-
-from sqlalchemy.exc import IntegrityError
-from sqlalchemy.orm import joinedload, Query
 
 from timApp.answer.answer import Answer
 from timApp.answer.answers import get_points_by_rule, get_latest_valid_answers_query
@@ -24,15 +13,9 @@ from timApp.timdb.sqa import db
 from timApp.user.user import User
 from timApp.user.usergroup import UserGroup
 from timApp.user.usergroupmember import UserGroupMember, membership_current
-from timApp.util.flask.requesthelper import RouteException
 from timApp.util.utils import get_current_time
 
 
-class PeerReviewException(Exception):
-    pass
-
-
-# Generate reviews groups for list
 def generate_review_groups(doc: DocInfo, tasks: list[Plugin]) -> None:
     task_ids = []
 
@@ -127,6 +110,10 @@ def generate_review_groups(doc: DocInfo, tasks: list[Plugin]) -> None:
     db.session.commit()
 
 
+class PeerReviewException(Exception):
+    pass
+
+
 def save_review(
     answer: Answer,
     task_id: TaskId,
@@ -160,100 +147,3 @@ def save_review(
 
     db.session.add(review)
     return review
-
-
-def get_reviews_for_user(d: DocInfo, user: User) -> list[PeerReview]:
-    q = get_reviews_for_user_query(d, user).options(joinedload(PeerReview.reviewable))
-    return q.all()
-
-
-def get_reviews_for_user_query(d: DocInfo, user: User) -> Query:
-    return PeerReview.query.filter_by(block_id=d.id, reviewer_id=user.id)
-
-
-def get_all_reviews(doc: DocInfo) -> list[PeerReview]:
-    return PeerReview.query.filter_by(block_id=doc.id).all()
-
-
-def get_reviews_to_user(d: DocInfo, user: User) -> list[PeerReview]:
-    q = get_reviews_to_user_query(d, user).options(joinedload(PeerReview.reviewable))
-    return q.all()
-
-
-def get_reviews_to_user_query(d: DocInfo, user: User) -> Query:
-    return PeerReview.query.filter_by(block_id=d.id, reviewable_id=user.id)
-
-
-def has_review_access(
-    doc: DocInfo,
-    reviewer_user: User,
-    task_id: TaskId | None,
-    reviewable_user: User | None,
-) -> bool:
-    if not is_peerreview_enabled(doc):
-        return False
-    q = PeerReview.query.filter_by(
-        block_id=doc.id,
-        reviewer_id=reviewer_user.id,
-    )
-    if task_id is not None:
-        q = q.filter_by(task_name=task_id.task_name)
-    if reviewable_user is not None:
-        q = q.filter_by(reviewable_id=reviewable_user.id)
-    return bool(q.first())
-
-
-def check_review_grouping(doc: DocInfo) -> bool:
-    q = PeerReview.query.filter_by(block_id=doc.id)
-    return bool(q.first())
-
-
-def is_peerreview_enabled(doc: DocInfo) -> bool:
-    settings = doc.document.get_settings()
-    peer_review_start = settings.peer_review_start()
-    peer_review_stop = settings.peer_review_stop()
-
-    if not peer_review_start or not peer_review_stop:
-        return doc.document.get_settings().peer_review()
-
-    return peer_review_start <= get_current_time() < peer_review_stop
-
-
-def get_reviews_for_document(doc: DocInfo) -> list[PeerReview]:
-    """Get peer-reviewers of current document from the database.
-    :param doc: Document containing reviewable answers.
-    """
-    return PeerReview.query.filter_by(
-        block_id=doc.id,
-    ).all()
-
-
-def change_peerreviewers_for_user(
-    doc: DocInfo,
-    task: str,
-    reviewable: int,
-    old_reviewers: list[int],
-    new_reviewers: list[int],
-) -> None:
-    """Change reviewers in one task.
-    :param doc: Document containing reviewable answers.
-    :param task: task name.
-    :param reviewable: User ID for the review target user.
-    :param old_reviewers: List of old reviewers IDs
-    :param new_reviewers: List of new reviewers IDs
-    """
-    # TODO: Clean up; don't do one query per loop: instead fetch all users at once!
-    for i in range(0, len(new_reviewers)):
-        updated_user = PeerReview.query.filter_by(
-            block_id=doc.id,
-            reviewer_id=old_reviewers[i],
-            reviewable_id=reviewable,
-            task_name=task,
-        ).first()
-        updated_user.reviewer_id = new_reviewers[i]
-    try:
-        db.session.flush()
-    except IntegrityError:
-        db.session.rollback()
-        raise RouteException("Error: Same reviewer more than once in one task")
-    # Note: Don't commit here because the function is used in other functions that update the DB
